phase1_complete/requirements.txt
# Phase 1完全版システム - 必須パッケージ
openai==1.6.1
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
Pillow==10.1.0
pillow-heif==0.13.0
PyPDF2==3.0.1
pdf2image==1.16.3
python-docx==1.1.0
openpyxl==3.1.2
beautifulsoup4==4.12.2
pytesseract==0.3.10
python-magic==0.4.27
mutagen==1.47.0
python-dotenv==1.0.0
